+++ src/PaymentMethod/KlarnaBankTransfer.php
@@ -89,4 +89,10 @@ class KlarnaBankTransfer extends KlarnaAbstract
      * @var bool
      */
     protected $additional_data_required = true;
+
+    /**
+     * Defines if this payment method should be hidden from the checkout or listing
+     * @var bool
+     */
+    protected $is_hidden = true;
 }