+++ templates/admin_refund_failed/text.php
@@ -5,6 +5,7 @@
 /** @var $shop_url */
 /** @var $customer_name */
 /** @var $order_reference */
+/** @var $platform_name */
 ?>
 
 <?php echo $view->__('admin_refund_failed.hello', [], 'email'); ?>
@@ -12,4 +13,4 @@
 <?php echo $view->__('admin_refund_failed.text2', [], 'email'); ?>
 <?php echo $view->__('admin_refund_failed.text3', [], 'email'); ?>
 <?php echo $view->__('admin_refund_failed.regards', [], 'email'); ?>,
-<?php echo $view->__('admin_refund_failed.team', [], 'email'); ?>
+<?php echo $view->__('admin_refund_failed.team', ['%platform_name%' => $platform_name], 'email'); ?>