+++ templates/onboarding_request/html.php
@@ -6,9 +6,10 @@
 /** @var $country */
 /** @var DateTime $requestTimeDate*/
 /** @var $versionNumber */
+/** @var $platform_name */
 ?>
 <?php echo $view->__('onboarding_request.dear', [], 'email'); ?>,<br />
-<?php echo $view->__('onboarding_request.text1', [], 'email'); ?> <?php echo $eCommercePlatform; ?>.<br /><br />
+<?php echo $view->__('onboarding_request.text1', ['%platform_name%' => $platform_name], 'email'); ?> <?php echo $eCommercePlatform; ?>.<br /><br />
 
 <strong><?php echo $view->__('onboarding_request.text2', [], 'email'); ?>:</strong><br />
 <?php echo $view->__('onboarding_request.text3', [], 'email'); ?>: <?php echo $companyName; ?><br />
@@ -20,5 +21,5 @@
 <?php echo $view->__('onboarding_request.text8', [], 'email'); ?>: <?php echo $versionNumber; ?><br /><br />
 
 <?php echo $view->__('onboarding_request.text9', [], 'email'); ?><br />
-<?php echo $view->__('onboarding_request.text10', [], 'email'); ?>,<br />
+<?php echo $platform_name; ?>,<br />
 <?php echo $view->__('onboarding_request.text11', [], 'email'); ?><br />